+++ src/louds/join_ranked_results.cpp
@@ -134,9 +134,9 @@ bool AND_ranked(qdag *Q[], uint16_t nQ,
             path += tupleQdags.path; // add the bits to the bitvector
             // compute the coordinates if it's a leaf
             if(cur_level == max_level){
-                uint32_t coordinates[nAtt];
+                uint16_t coordinates[nAtt];
                 delete[] root_temp;
-                for(uint32_t k = 0; k < nAtt; k++){
+                for(uint16_t k = 0; k < nAtt; k++){
                     coordinates[k] = 0;
                 }
                 getCoordinates(path, l, max_level, coordinates);
